--- docker configs/docker_selenium/dockerfile
@@ -1,22 +1,0 @@
-# modified from https://github.com/kevinhowbrook/Selenium-Python-Docker-scraper/blob/master/Dockerfile
-
-FROM python:latest
-#  install required linux packages
-RUN apt-get update \
-    && apt-get -y install gcc make \
-    && rm -rf /var/lib/apt/lists/*s
-# install google chrome
-RUN wget -q -O - https://dl-ssl.google.com/linux/linux_signing_key.pub | apt-key add -
-RUN sh -c 'echo "deb [arch=amd64] http://dl.google.com/linux/chrome/deb/ stable main" >> /etc/apt/sources.list.d/google-chrome.list'
-RUN apt-get -y update; \
-	apt-get install xvfb -y; \
-	apt-get install -y google-chrome-stable
-# run pip required commands
-
-RUN pip install --no-cache-dir --upgrade pip
-WORKDIR /python_scrape
-COPY ./requirements.txt /python_scrape
-RUN pip install --no-cache-dir -r requirements.txt
-
-# cause a loop so docker does not close - disabled for now.
-CMD tail -f /dev/null
