+++ docs/postgresql/examples/test_schema_create.sh
