+++ src/pgrest_conf/db_params.json
@@ -1,0 +1,4 @@
+{
+  "role": "data_user",
+  "jwt_secret": "i564XNS9Lj4YF7EapvBRYV1OKxSDPGFl"
+}
